requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
flask-migrate==4.1.0
flask-jwt-extended==4.6.0
flask-cors==5.0.0
cloudinary==1.44.1
pytest==7.4.4
pytest-flask==1.3.0
python-dotenv==1.0.1
requests==2.31.0